Adding more details
adding files and distributing functions to pass norminette
