+++ src/minishell.c
@@ -6,295 +6,11 @@
 /*   By: mac <mac@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/12 17:52:30 by vics              #+#    #+#             */
-/*   Updated: 2024/02/10 16:11:29 by mac              ###   ########.fr       */
+/*   Updated: 2024/02/22 11:13:53 by mac              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
-
-char *save_sentence_l(char *string, int num)
-{
-    int i = 0;
-    int j = 0;
-    char *sentence = calloc((num + 100), sizeof(char));
-    char *final;
-    while (valid_chars(string[i]) || string[i] == ' ')
-        i++;
-    while (i < num && string[i] != '\0') {
-        sentence[j] = string[i];
-        i++;
-        j++;
-    }
-    sentence[i] = '\0';
-    final = ft_strtrim(sentence, " \t\n");
-    free(sentence);
-    return (final);
-}
-
-char *save_sentence_r(char *string, int num)
-{
-    int i = num;
-    int j = 0;
-    char *sentence = calloc((strlen(string) - num + 1), sizeof(char));
-    char *final;
-    while (string[i] && valid_chars(string[i]))
-        i++;
-    while (string[i] != '\0') {
-        sentence[j] = string[i];
-        i++;
-        j++;
-    }
-    sentence[j] = '\0';
-    final = ft_strtrim(sentence, " \t\n");
-    free(sentence);
-    return (final);
-}
-
-void    exec_error(char *command, char *error_msg)
-{
-    int fd;
-    //char *s;
-    
-    fd = dup(1);
-    dup2(2, 1);
-    printf("%s: %s", command, error_msg);
-    dup2(fd, 1);
-}
-
-void    stx_error(char *error_msg)
-{
-    int fd;
-    //char *s;
-    
-    fd = dup(1);
-    dup2(2, 1);
-    printf("%s", error_msg);
-    dup2(fd, 1);
-}
-
-void    stx_error_op(char *error_msg, char op)
-{
-    int fd;
-    //char *s;
-    
-    fd = dup(1);
-    dup2(2, 1);
-    printf("%s '%c'\n", error_msg, op);
-    dup2(fd, 1);
-}
-
-void    check_operator(t_info_tree *tree)
-{
-    if (ft_strlen(tree->operator) > 2)
-        stx_error_op(WRONG_OP, tree->operator[0]);
-    else if (tree->operator != NULL && ((!tree->left->command[0] && ft_strcmp(tree->operator, "<<") != 0) || !tree->right->command[0]))
-        stx_error(WRONG_OP_NL);    
-}
-
-bool	ft_replace_chrchr(char *str, char find, char replace)
-{
-	int		i;
-	bool	n;
-
-	i = 0;
-	n = false;
-	while (str[i])
-	{
-		if (str[i] == find)
-		{
-			str[i] = replace;
-			n = true;
-		}
-		i++;
-	}
-	return (n);
-}
-
-void	ft_remove_chr(char *str, char find)
-{
-	int		i;
-    int     j;
-	bool	n;
-
-	i = 0;
-    j = 0;
-	n = false;
-	while (str[i])
-	{
-		if (str[i] == find)
-			i++;
-        else
-        {
-            str[j] = str[i];
-		    j++;
-            i++;
-        }
-	}
-    str[j] = '\0';
-}
-
-char *ft_replace_strstr(char *string, int index, int len, char *replace)
-{
-    int j = 0;
-    int i = 0;
-
-    char *new = malloc(sizeof(char) * (ft_strlen(string) + ft_strlen(replace) - len + 1));
-    while (string[i] && i < index)
-    {
-        new[j] = string[i];
-        j++;
-        i++;
-    }
-    i += len + 1;
-    if (replace && replace[0])
-    {
-        while (*replace)
-            new[j++] = *replace++;
-    }
-    while (string[i])
-        new[j++] = string[i++];
-    new[j] = '\0';
-    free(string);
-    return(new);
-}
-
-char *expansion(t_var *var, char *command)
-{
-    int i = 0;
-    int len = 0;
-    int last = 0;
-    char *name = NULL;
-    char *value_str = NULL;
-    t_env *env;
-    
-    while (last >= 0)
-    {
-        i = ft_strchr_index(&command[last], '$');
-        if (i != -1)
-        {
-            i += last + 1;
-            while (command[i] && ft_isalnum(command[i]))
-            {
-                i++;
-                len++;
-            }
-        }
-        if (len == 0)
-            last = -1;
-        else
-        {
-            name = malloc(sizeof(char) * len + 1);
-            ft_strlcpy(name, &command[i - len], len + 1);
-            env = find_in_env(var->env, name);
-            free(name);
-            if (env)
-                value_str = ft_strdup(env->value);
-            else if (!ft_isalpha(command[ft_strchr_index(&command[last], '$') + 1]))
-            {
-                int index = ft_strchr_index(&command[last], '$') + 2;
-                value_str = ft_substr(command, index, len - 1);
-            }
-            command = ft_replace_strstr(command, i - len - 1, len, value_str);
-            last = i - ft_strlen(value_str);
-            free(value_str);
-            value_str = "";
-        }
-        len = 0;
-    }
-    return(command);
-}
-
-void recursive_tree(t_var *var, t_info_tree *tree, char *string)
-{
-    int j = get_biggest_priority(string);
-
-    if (j == -1) {
-        tree->left = NULL;
-        tree->right = NULL;
-        tree->operator = NULL;
-        tree->command = string;
-        ft_remove_chr(tree->command, '"');
-        ft_remove_chr(tree->command, '\'');
-        
-    } else {
-        tree->command = NULL;
-        tree->operator = get_operator(string, j);
-        tree->left = init_linked_tree(save_sentence_l(string, j), tree->operator, tree->operator);
-        tree->right = init_linked_tree(save_sentence_r(string, j), tree->operator, tree->operator);
-        check_operator(tree);
-        ft_remove_chr(tree->left->command, '"');
-        ft_remove_chr(tree->left->command, '\'');
-        ft_remove_chr(tree->right->command, '"');
-        ft_remove_chr(tree->right->command, '\'');
-        recursive_tree(var, tree->left, tree->left->command);
-        recursive_tree(var, tree->right, tree->right->command);
-    }
-}
-
-void recursive2(t_var *var, t_info_tree *tree)
-{
-    int copia_0 = dup(0);
-    int copia_1 = dup(1);
-
-    if (tree->operator != NULL)
-        function_ptr_op(var, tree);
-    if (tree->left != NULL && tree->left->left != NULL)
-        recursive2(var, tree->left);
-    if (tree->operator == NULL)
-        func_pipe(var, tree->command);
-    if (tree->right != NULL && tree->right->left != NULL)
-        recursive2(var, tree->right);
-    dup2(copia_0, 0);
-    dup2(copia_1, 1);
-}
-
-void make_binnary_tree(t_var *var, char *line)
-{
-    t_info_tree *tmp = init_struct_tree();
-	
-    var->tree = tmp;
-    line = expansion(var, line);
-    recursive_tree(var, tmp, line);
-    tmp = var->tree;
-    recursive2(var, tmp);
-    tmp = var->tree;
-    free_binnarytree(tmp);
-}
-
-void free_lst(t_env* head) {
-    t_env* current = head;
-    t_env* next;
-
-    while (current != NULL) {
-        next = current->next;
-        free(current->name);
-        free(current->value);
-        free(current);
-        current = next;
-    }
-}
-
-void free_binnarytree(t_info_tree *tree)
-{
-    if (!tree)
-        return;
-    free_binnarytree(tree->left);
-    free_binnarytree(tree->right);
-
-    if (tree->command)
-        free(tree->command);
-    if (tree->operator)
-        free(tree->operator);
-    free(tree);
-}
-
-void    func_exit(t_var *var)
-{
-    free_lst(var->env);
-    free(var->act);
-    free(var->op);
-    free(var);
-}
 
 int main(int argc, char **argv, char **env) {
     char *line = NULL;
@@ -305,7 +21,6 @@ int main(int argc, char **argv, char **env) {
     (void)argv;
     t_var *var = init_struct(env);
     previous_str = NULL;
-    init_signals();
    
 	 while(1)
 	{
@@ -313,6 +28,7 @@ int main(int argc, char **argv, char **env) {
 		line = readline(path);
         line_cleaned = NULL;
         free(path);
+        init_signals();
         if (line && line[0] != '\0')
         {
 		    line_cleaned = ft_strtrim(line, " \t\n");
