+++ src/signals.c
@@ -4,19 +4,32 @@ void	sigintHandler(int signus)
 {
     (void)signus;
     printf ("\n Ctrl-C: Deteniendo la ejecución. Volviendo al shell\n");
-    exit(EXIT_SUCCESS);
 }
 
 void	sigquitHandler(int signus)
 {
     (void)signus;
     printf ("\n Ctrl-\\: Interrumpiendo la ejecución. Volviendo al shell\n");
-    exit(EXIT_SUCCESS);
 }
 
 void    init_signals(void)
 {
-    signal(SIGINT, sigintHandler);
-    signal(SIGQUIT, sigquitHandler);
+    struct sigaction    sigact;
 
+    sigact.sa_handler = sigintHandler;
+    sigemptyset(&sigact.sa_mask);
+    sigact.sa_flags = 0;
+
+    if (sigaction(SIGINT, &sigact, NULL) < 0)
+    {
+        perror("sigaction");
+        exit(EXIT_FAILURE);
+    }
+    sigact.sa_handler = sigquitHandler;
+    if (sigaction(SIGQUIT, &sigact, NULL) < 0)
+    {
+        perror("sigaction");
+        exit(EXIT_FAILURE);
+    }
+    
 }
